Extract IPC handlers and menu builder out of main/index.ts
registerIpc() previously inlined 27 handlers spanning projects, trees, repos, and settings raw read/write — the largest single contributor to main/index.ts being 909 LOC. Split along the same shape as the existing ipc/{settings,system,terminal}.ts modules:

- ipc/projects.ts — listProjects, getProject, step.*, listProjectFiles, setStatus (with the lingering-branch warning), createProject, note.*, settings.read/writeRaw, project.createNote. assertUnderConception lives here too since every caller is in this module.
- ipc/trees.ts — read{Knowledge,Resources,Skills}Tree, tree.*, search.
- ipc/repos.ts — listRepos, listReposForPrimary, invalidateGitStatus, getDirtyDetails, listOpenWith, launchOpenWith, forceStopRepo. Watcher reconciliation stays inline with the listRepos verbs.
- main/menu.ts — buildMenu/rebuildMenu/rebuildMenuFromSettings + the prettyRecentLabel helper. Window reference is injected via setMenuWindow so menu callbacks fire IPC at the active window.

registerIpc() collapses to a flat list of register*Ipc() calls. Behaviour-preserving — handler bodies copied verbatim, only the ambient closure (assertUnderConception) is converted to an in-module helper. main/index.ts: 909 -> 329 LOC.

Updates docs/explanation/internals.md "IPC contract" section to point at the per-domain modules.
