+++ src/main/index.ts
@@ -1,51 +1,19 @@
-import { app, BrowserWindow, ipcMain, Menu, net, protocol, shell } from 'electron';
-import type { MenuItemConstructorOptions } from 'electron';
+import { app, BrowserWindow, net, protocol, shell } from 'electron';
 import { promises as fsp } from 'node:fs';
 import { join } from 'node:path';
 import { pathToFileURL } from 'node:url';
-import { parseHeader } from '../shared/header';
 
-import { DEFAULT_LAYOUT, readSettings, settingsPath } from './settings';
-import { findProjectReadmes } from './walk';
-import { parseReadme } from './parse';
+import { DEFAULT_LAYOUT, readSettings } from './settings';
 import { setWatchedConception } from './watcher';
-import { addStep, editStepText, transitionStatus, toggleStep, writeNote } from './mutate';
-import { touchDirtyMarker } from './dirty';
-import { createProjectCore } from './create-project';
-import { checkBranchState } from './worktree-ops';
-import { listProjectFiles } from './files';
-import { requirePathUnder, requirePathUnderWorkspace } from './path-bounds';
-import { readKnowledgeTree } from './knowledge';
-import { readResourcesTree } from './resources';
-import { readSkillsTree } from './skills';
-import { treeCreateMd, treeImportFile, treeMkdir } from './tree-mutations';
-import { resolveConceptionPaths } from './conception-paths';
-import { createProjectNote, readNote } from './note';
-import { search } from './search';
-import { listRepos, listReposForPrimary } from './repos';
-import { invalidateAll } from './git-status-cache';
-import {
-  disposeRepoWatchers,
-  recomputeAllWatchedRepos,
-  setRepoWatchers,
-  watchTargetsFromRepos,
-} from './repo-watchers';
-import { getDirtyDetails } from './git-details';
-import { forceStopRepo, launchOpenWith, listOpenWith } from './launchers';
+import { disposeRepoWatchers } from './repo-watchers';
 import { killAll, migrateTerminalFromConfigIfNeeded } from './terminals';
-import { registerTerminalIpc } from './ipc/terminal';
+import { buildMenu, rebuildMenu, rebuildMenuFromSettings, setMenuWindow } from './menu';
+import { registerProjectsIpc } from './ipc/projects';
+import { registerReposIpc } from './ipc/repos';
 import { registerSettingsIpc } from './ipc/settings';
 import { registerSystemIpc } from './ipc/system';
-import type {
-  LayoutState,
-  OpenWithSlotKey,
-  Project,
-  ProjectCreateInput,
-  ProjectCreateResult,
-  StepMarker,
-  TreeRoot,
-} from '../shared/types';
-import { statusOrder } from '../shared/projects';
+import { registerTerminalIpc } from './ipc/terminal';
+import { registerTreesIpc } from './ipc/trees';
 
 // Hard flip from v2.14.0: CLI nouns belong on `condash-cli`, not `condash`.
 // If a user types a CLI noun on `condash`, error before any GUI init so the
@@ -220,314 +188,6 @@ async function createWindow(initialPath: string | null): Promise<BrowserWindow> 
 }
 
 /**
- * Build the application menu. The View submenu mirrors the unified
- * layout's pane-visibility state — Show/Hide Projects + Show/Hide
- * Terminal as toggles, plus a three-state group (Code | Knowledge |
- * neither) for the right-slot working surface. Pass the current layout
- * so check marks line up with what's actually shown; rebuild the menu
- * after any layout change so the marks refresh. No Quit accelerator on
- * purpose: Ctrl+Q is too easy to hit by accident, and File → Quit
- * routes through a renderer-side confirmation modal anyway.
- */
-function buildMenu(
-  layout: LayoutState = DEFAULT_LAYOUT,
-  recents: { paths: string[]; current: string | null } = { paths: [], current: null },
-): void {
-  const send = (command: string): void => {
-    mainWindow?.webContents.send('menu-command', command);
-  };
-
-  const openRecentSubmenu: MenuItemConstructorOptions[] = recents.paths.length
-    ? [
-        ...recents.paths.map((path) => ({
-          label: prettyRecentLabel(path),
-          type: 'checkbox' as const,
-          checked: path === recents.current,
-          click: () => {
-            mainWindow?.webContents.send('menu-open-recent', path);
-          },
-        })),
-        { type: 'separator' as const },
-        {
-          label: 'Clear menu',
-          click: () => {
-            mainWindow?.webContents.send('menu-clear-recents');
-          },
-        },
-      ]
-    : [{ label: '(no recent conceptions)', enabled: false }];
-
-  const fileSubmenu: MenuItemConstructorOptions[] = [
-    {
-      label: 'Open…',
-      accelerator: 'CommandOrControl+O',
-      click: () => send('open-folder'),
-    },
-    {
-      label: 'Open Recent',
-      submenu: openRecentSubmenu,
-    },
-    {
-      label: 'Open conception directory',
-      click: () => send('open-conception'),
-    },
-    {
-      label: 'New project…',
-      accelerator: 'CommandOrControl+N',
-      click: () => send('new-project'),
-    },
-    { type: 'separator' },
-    {
-      label: 'Settings',
-      accelerator: 'CommandOrControl+,',
-      click: () => send('open-settings'),
-    },
-    {
-      // Two accelerators for the same action — Electron menus only honour
-      // one accelerator per item, so the Ctrl+K binding is wired in the
-      // renderer's handleGlobalKeyDown (the cheat-sheet documents both).
-      label: 'Search…',
-      accelerator: 'CommandOrControl+Shift+F',
-      click: () => send('search'),
-    },
-    { type: 'separator' },
-    {
-      label: 'Quit',
-      // No accelerator on purpose — see the comment above buildMenu().
-      click: () => send('request-quit'),
-    },
-  ];
-
-  const viewSubmenu: MenuItemConstructorOptions[] = [
-    {
-      label: 'Show Projects',
-      type: 'checkbox',
-      checked: layout.projects,
-      click: () => send('toggle-projects'),
-    },
-    {
-      label: 'Show Code',
-      type: 'checkbox',
-      checked: layout.working === 'code',
-      accelerator: 'CommandOrControl+Shift+C',
-      click: () => send('show-code'),
-    },
-    {
-      label: 'Show Knowledge',
-      type: 'checkbox',
-      checked: layout.working === 'knowledge',
-      accelerator: 'CommandOrControl+Shift+K',
-      click: () => send('show-knowledge'),
-    },
-    {
-      label: 'Show Resources',
-      type: 'checkbox',
-      checked: layout.working === 'resources',
-      accelerator: 'CommandOrControl+R',
-      click: () => send('show-resources'),
-    },
-    {
-      label: 'Show Skills',
-      type: 'checkbox',
-      checked: layout.working === 'skills',
-      accelerator: 'CommandOrControl+L',
-      click: () => send('show-skills'),
-    },
-    {
-      label: 'Hide working surface',
-      enabled: layout.working !== null,
-      click: () => send('hide-working'),
-    },
-    {
-      label: 'Show Terminal',
-      type: 'checkbox',
-      checked: layout.terminal,
-      accelerator: 'CommandOrControl+`',
-      click: () => send('toggle-terminal'),
-    },
-    { type: 'separator' },
-    {
-      label: 'Refresh',
-      accelerator: 'F5',
-      click: () => send('refresh'),
-    },
-    { type: 'separator' },
-    {
-      role: 'reload',
-      label: 'Reload window',
-      // Reload normally lives at Ctrl+R, but we hand that accelerator to the
-      // Resources panel (more useful in day-to-day work). Reload moves to
-      // Ctrl+Shift+R, which still aligns with browser muscle memory for
-      // a hard reload.
-      accelerator: 'CommandOrControl+Shift+R',
-    },
-    { role: 'toggleDevTools' },
-    { type: 'separator' },
-    { role: 'resetZoom' },
-    { role: 'zoomIn' },
-    { role: 'zoomOut' },
-    { type: 'separator' },
-    { role: 'togglefullscreen' },
-  ];
-
-  const helpSubmenu: MenuItemConstructorOptions[] = [
-    {
-      label: 'About Condash',
-      click: () => send('about'),
-    },
-    { type: 'separator' },
-    { label: 'Welcome', click: () => send('help-welcome') },
-    { label: 'Quick start', click: () => send('help-quick-start') },
-    { label: 'Keyboard shortcuts', click: () => send('help-shortcuts') },
-    { type: 'separator' },
-    { label: 'Configuration', click: () => send('help-configuration') },
-    { label: 'CLI overview', click: () => send('help-cli') },
-    { label: 'Why Markdown-first', click: () => send('help-why-markdown') },
-    { type: 'separator' },
-    {
-      label: 'Open documentation site',
-      click: () => {
-        void shell
-          .openExternal('https://condash.vcoeur.com')
-          .catch((err) => console.error('[menu] openExternal failed', err));
-      },
-    },
-    {
-      label: 'Open issue tracker',
-      click: () => {
-        void shell
-          .openExternal('https://github.com/vcoeur/condash/issues')
-          .catch((err) => console.error('[menu] openExternal failed', err));
-      },
-    },
-  ];
-
-  const template: MenuItemConstructorOptions[] = [
-    { label: 'File', submenu: fileSubmenu },
-    {
-      label: 'Edit',
-      submenu: [
-        { role: 'undo' },
-        { role: 'redo' },
-        { type: 'separator' },
-        { role: 'cut' },
-        { role: 'copy' },
-        { role: 'paste' },
-        { role: 'selectAll' },
-      ],
-    },
-    { label: 'View', submenu: viewSubmenu },
-    { label: 'Help', role: 'help', submenu: helpSubmenu },
-  ];
-
-  Menu.setApplicationMenu(Menu.buildFromTemplate(template));
-}
-
-/**
- * Rebuild the application menu against the latest layout. Used as the
- * `onLayoutChange` hook by registerSettingsIpc — the View submenu's check
- * marks need to refresh after every layout mutation.
- */
-let lastRecents: { paths: string[]; current: string | null } = { paths: [], current: null };
-let lastLayout: LayoutState = DEFAULT_LAYOUT;
-function rebuildMenu(layout: LayoutState): void {
-  lastLayout = layout;
-  buildMenu(layout, lastRecents);
-}
-async function rebuildMenuFromSettings(): Promise<void> {
-  const settings = await readSettings();
-  lastRecents = {
-    paths: settings.recentConceptionPaths ?? [],
-    current: settings.lastConceptionPath,
-  };
-  buildMenu(lastLayout, lastRecents);
-}
-
-/**
- * Format a recents path for the File → Open Recent submenu. The basename
- * (e.g. `conception`) carries the visual weight; the parent directory is
- * shown after a dimmed em-dash so two trees with the same basename don't
- * look identical.
- */
-function prettyRecentLabel(path: string): string {
-  const segments = path.split('/').filter((s) => s.length > 0);
-  if (segments.length === 0) return path;
-  const tail = segments[segments.length - 1];
-  const head = segments.slice(0, -1).join('/');
-  if (!head) return tail;
-  // The Electron menu strips raw `/` runs in some renders; collapse the
-  // parent path to a tilde-anchored form when it lives under HOME.
-  return `${tail} — ${head}`;
-}
-
-/**
- * Defence-in-depth: every IPC handler that accepts a `path` from the
- * renderer and reads or writes the filesystem at that path runs through
- * here first. Realpathed bound check against the current conception
- * root via shared/path-bounds — pass-4..6 deferred this sweep, pass-7
- * lands the conception-bound subset (note read/write, step ops,
- * getProject, listProjectFiles, project.createNote). Repos /
- * worktrees / screenshot_dir handlers stay open for now because their
- * threat model targets a different bound.
- */
-async function assertUnderConception(path: string): Promise<void> {
-  const { lastConceptionPath: conceptionPath } = await readSettings();
-  if (!conceptionPath) {
-    throw new Error('no conception path is set');
-  }
-  await requirePathUnder(path, conceptionPath);
-}
-
-async function listProjects(): Promise<Project[]> {
-  const { lastConceptionPath: conceptionPath } = await readSettings();
-  if (!conceptionPath) return [];
-
-  const readmes = await findProjectReadmes(conceptionPath);
-  const projects = await Promise.all(readmes.map(parseReadme));
-
-  return projects.sort((a, b) => {
-    const o = statusOrder(a.status) - statusOrder(b.status);
-    if (o !== 0) return o;
-    return a.slug.localeCompare(b.slug);
-  });
-}
-
-async function readBranchFromReadme(readmePath: string): Promise<string | null> {
-  try {
-    const raw = await fsp.readFile(readmePath, 'utf8');
-    return parseHeader(raw).branch;
-  } catch {
-    return null;
-  }
-}
-
-function buildBranchWarning(
-  branch: string,
-  lingeringWorktrees: { expectedWorktree: string }[],
-  lingeringBranches: { name: string }[],
-): string {
-  const parts: string[] = [];
-  if (lingeringWorktrees.length > 0) {
-    const paths = lingeringWorktrees.map((r) => r.expectedWorktree).join(', ');
-    parts.push(`worktree(s) still on disk at ${paths}`);
-  }
-  if (lingeringBranches.length > 0) {
-    const repos = lingeringBranches.map((r) => r.name).join(', ');
-    parts.push(`local branch '${branch}' still exists in ${repos}`);
-  }
-  return `${parts.join('; ')} — run \`condash-cli worktrees remove ${branch}\` then \`git branch -d ${branch}\` to clean up.`;
-}
-
-async function getProject(path: string): Promise<Project | null> {
-  try {
-    return await parseReadme(path);
-  } catch (err) {
-    if ((err as NodeJS.ErrnoException).code === 'ENOENT') return null;
-    throw err;
-  }
-}
-
-/**
  * Serve files referenced from note-view markdown (images via relative path)
  * over a `condash-file:///abs/path` URL. Sandboxed: the requested path must
  * resolve under the current conception path. The renderer rewrites relative
@@ -581,128 +241,27 @@ function isPathUnder(child: string, parent: string): boolean {
   return c === p || c.startsWith(p);
 }
 
+/**
+ * Wire every IPC handler. Per-domain modules under `ipc/` own the actual
+ * handler bodies; this dispatcher just calls each module's
+ * `register*Ipc(opts?)` once at app start. New verbs go in (or alongside)
+ * the matching domain module — keep this body a flat list of registrations.
+ *
+ * The per-domain split:
+ *   • projects — listProjects, getProject, step.*, listProjectFiles,
+ *     setStatus, createProject, note.*, settings.read/writeRaw,
+ *     project.createNote.
+ *   • trees — read{Knowledge,Resources,Skills}Tree, tree.*, search.
+ *   • repos — listRepos, listReposForPrimary, invalidateGitStatus,
+ *     getDirtyDetails, listOpenWith, launchOpenWith, forceStopRepo.
+ *   • settings — theme/layout/welcome/cardMinWidth/treeExpansion/getSettingsPath.
+ *   • system — shell-out + app-info + conception path mutation.
+ *   • terminal — every term.* verb.
+ */
 function registerIpc(): void {
-  ipcMain.handle('listProjects', () => listProjects());
-
-  ipcMain.handle('getProject', async (_, path: string) => {
-    await assertUnderConception(path);
-    return getProject(path);
-  });
-
-  ipcMain.handle('readKnowledgeTree', async () => {
-    const { lastConceptionPath: conceptionPath } = await readSettings();
-    if (!conceptionPath) return null;
-    return readKnowledgeTree(conceptionPath);
-  });
-
-  ipcMain.handle('readResourcesTree', async () => {
-    const { lastConceptionPath: conceptionPath } = await readSettings();
-    if (!conceptionPath) return null;
-    const { resources } = await resolveConceptionPaths(conceptionPath);
-    return readResourcesTree(conceptionPath, resources);
-  });
-
-  ipcMain.handle('readSkillsTree', async () => {
-    const { lastConceptionPath: conceptionPath } = await readSettings();
-    if (!conceptionPath) return null;
-    const { skills } = await resolveConceptionPaths(conceptionPath);
-    return readSkillsTree(conceptionPath, skills);
-  });
-
-  ipcMain.handle('tree.createMd', (_, root: TreeRoot, dirRelPath: string, filename: string) =>
-    treeCreateMd(root, dirRelPath, filename),
-  );
-
-  ipcMain.handle('tree.mkdir', (_, root: TreeRoot, dirRelPath: string, name: string) =>
-    treeMkdir(root, dirRelPath, name),
-  );
-
-  ipcMain.handle('tree.importFile', (_, root: TreeRoot, dirRelPath: string) =>
-    treeImportFile(root, dirRelPath),
-  );
-
-  ipcMain.handle('search', async (_, query: string) => {
-    const { lastConceptionPath: conceptionPath } = await readSettings();
-    if (!conceptionPath) return [];
-    return search(conceptionPath, query);
-  });
-
-  ipcMain.handle('listRepos', async () => {
-    const { lastConceptionPath: conceptionPath } = await readSettings();
-    if (!conceptionPath) return [];
-    const repos = await listRepos(conceptionPath);
-    // Sync the per-repo FS watchers to the live repo set: a config edit
-    // that adds or removes a repo is reflected here, since this handler
-    // re-runs on every renderer-driven repos refresh.
-    await setRepoWatchers(watchTargetsFromRepos(repos));
-    return repos;
-  });
-
-  // Per-primary partial reload — driven by the structural FS watcher when
-  // `.git/HEAD` or `.git/worktrees/` changes. Returns the primary's
-  // RepoEntry plus its submodule children, freshly re-read. Watchers are
-  // re-synced for the affected paths so a freshly-added worktree gets a
-  // scalar watcher pair right away (and a freshly-removed one is dropped).
-  ipcMain.handle('listReposForPrimary', async (_, primaryName: string) => {
-    const { lastConceptionPath: conceptionPath } = await readSettings();
-    if (!conceptionPath) return [];
-    const entries = await listReposForPrimary(conceptionPath, primaryName);
-    // Re-list the *full* watcher set: the simplest correct way to make sure
-    // an added or removed worktree under this primary is reflected in the
-    // watch set without diffing the per-primary subset against the global
-    // one. The cost is one extra `listRepos` call on a structural event,
-    // which is rare (worktree mutation, branch checkout) — far cheaper
-    // than getting the watch-set delta logic wrong.
-    const repos = await listRepos(conceptionPath);
-    await setRepoWatchers(watchTargetsFromRepos(repos));
-    return entries;
-  });
-
-  // Drop the per-worktree git status cache + force-recompute every watched
-  // repo and broadcast `repo-events`. Wired to the renderer's F5 / Refresh
-  // path so the user sees fresh counts without waiting for an FS event,
-  // and without the renderer needing to refetch the whole repo list (which
-  // would tear down dropdowns/popovers).
-  ipcMain.handle('invalidateGitStatus', async () => {
-    invalidateAll();
-    await recomputeAllWatchedRepos();
-  });
-
-  // Click-to-inspect on the per-branch dirty badge. Returns the parsed
-  // `git status` line set + a `git diff --stat HEAD` snippet so the user
-  // can see what's dirty without dropping into a shell. Bound to the
-  // workspace + worktrees roots so a compromised renderer can't drive a
-  // shell-out `git status` against an arbitrary directory on disk.
-  ipcMain.handle(
-    'getDirtyDetails',
-    async (_, path: string, opts?: { scopeToSubtree?: boolean }) => {
-      const realPath = await requirePathUnderWorkspace(path);
-      return getDirtyDetails(realPath, opts ?? {});
-    },
-  );
-
-  ipcMain.handle('listOpenWith', async () => {
-    const { lastConceptionPath: conceptionPath } = await readSettings();
-    if (!conceptionPath) return {};
-    return listOpenWith(conceptionPath);
-  });
-
-  ipcMain.handle('launchOpenWith', async (_, slot: OpenWithSlotKey, path: string) => {
-    const { lastConceptionPath: conceptionPath } = await readSettings();
-    if (!conceptionPath) throw new Error('No conception path set');
-    // Bound to workspace + worktrees + conception so the renderer can
-    // launch the user's IDE on a project README, a workspace repo, or a
-    // worktree — but not on `/etc/passwd` or `~/.ssh/`.
-    const realPath = await requirePathUnderWorkspace(path);
-    return launchOpenWith(conceptionPath, slot, realPath);
-  });
-
-  ipcMain.handle('forceStopRepo', async (_, repoName: string) => {
-    const { lastConceptionPath: conceptionPath } = await readSettings();
-    if (!conceptionPath) throw new Error('No conception path set');
-    return forceStopRepo(conceptionPath, repoName);
-  });
-
+  registerProjectsIpc();
+  registerTreesIpc();
+  registerReposIpc();
   registerTerminalIpc();
   registerSettingsIpc({ onLayoutChange: rebuildMenu });
   registerSystemIpc({
@@ -713,147 +272,6 @@ function registerIpc(): void {
     onRecentsChange: () => {
       void rebuildMenuFromSettings();
     },
-  });
-
-  ipcMain.handle(
-    'step.toggle',
-    async (
-      _,
-      path: string,
-      lineIndex: number,
-      expectedMarker: StepMarker,
-      newMarker: StepMarker,
-    ) => {
-      await assertUnderConception(path);
-      return toggleStep(path, lineIndex, expectedMarker, newMarker);
-    },
-  );
-
-  ipcMain.handle(
-    'step.editText',
-    async (_, path: string, lineIndex: number, expectedText: string, newText: string) => {
-      await assertUnderConception(path);
-      return editStepText(path, lineIndex, expectedText, newText);
-    },
-  );
-
-  ipcMain.handle('step.add', async (_, path: string, text: string) => {
-    await assertUnderConception(path);
-    return addStep(path, text);
-  });
-
-  ipcMain.handle('listProjectFiles', async (_, path: string) => {
-    await assertUnderConception(path);
-    return listProjectFiles(path);
-  });
-
-  ipcMain.handle(
-    'setStatus',
-    async (_, path: string, newStatus: string, opts?: { summary?: string }) => {
-      const result = await transitionStatus(path, newStatus, opts);
-      // Touch the dirty marker so a follow-up `condash-cli projects index` is
-      // surfaced. Best-effort: if the conception path isn't set we just
-      // skip it — the in-memory list rebuild still happens via the watcher.
-      const { lastConceptionPath: conceptionPath } = await readSettings();
-      if (conceptionPath) {
-        await touchDirtyMarker(conceptionPath, 'projects').catch((err) => {
-          console.error('[setStatus] touchDirtyMarker failed', err);
-        });
-        // On close (done-edge: prev !== done, new === done), surface any
-        // leftover-branch warnings so the GUI can toast them — silently
-        // swallowing the miss let the same broken cleanup ship twice in
-        // April. Keep the call best-effort so a failed probe never blocks
-        // the close itself.
-        if (result.timelineAppended && /^- .* — Closed/.test(result.timelineAppended)) {
-          const branch = await readBranchFromReadme(path);
-          if (branch) {
-            try {
-              const state = await checkBranchState(conceptionPath, branch);
-              const lingeringWorktrees = state.repos.filter((r) => r.worktreeExists);
-              const lingeringBranches = state.repos.filter((r) => r.localBranchExists);
-              if (lingeringWorktrees.length > 0 || lingeringBranches.length > 0) {
-                result.branchWarning = buildBranchWarning(
-                  branch,
-                  lingeringWorktrees,
-                  lingeringBranches,
-                );
-              }
-            } catch {
-              // Best-effort probe — never block the close on its own failure.
-            }
-          }
-        }
-      }
-      return result;
-    },
-  );
-
-  ipcMain.handle(
-    'createProject',
-    async (_, input: ProjectCreateInput): Promise<ProjectCreateResult> => {
-      const { lastConceptionPath: conceptionPath } = await readSettings();
-      if (!conceptionPath) throw new Error('No conception path set');
-      const result = await createProjectCore(conceptionPath, {
-        kind: input.kind,
-        slug: input.slug,
-        title: input.title,
-        // Apps stays empty for the GUI quick-create form; users fill it in
-        // by editing the README or via the popup.
-        apps: [],
-        branch: null,
-        base: null,
-        severity: input.severity ?? null,
-        severityImpact: input.severityImpact ?? null,
-        environment: input.environment ?? null,
-      });
-      // The renderer asked for status `now | review | later | backlog`, but
-      // createProjectCore always renders `Status: now`. If the user picked
-      // anything else, flip it now via the same transitionStatus primitive
-      // — that keeps the status-write path single-source.
-      if (input.status !== 'now') {
-        await transitionStatus(result.readme, input.status);
-      }
-      return {
-        slug: result.slug,
-        path: result.path,
-        relPath: result.relPath,
-        readme: result.readme,
-      };
-    },
-  );
-
-  ipcMain.handle('note.read', async (_, path: string) => {
-    await assertUnderConception(path);
-    return readNote(path);
-  });
-
-  ipcMain.handle(
-    'note.write',
-    async (_, path: string, expectedContent: string, newContent: string) => {
-      await assertUnderConception(path);
-      return writeNote(path, expectedContent, newContent);
-    },
-  );
-
-  // Raw read of the per-machine `settings.json`. Used by the Settings modal
-  // to compute inheritance badges (compare global vs. condash.json values)
-  // and to drive the Global-tab editor through the same patchConfig flow it
-  // uses for condash.json. Returns `''` when the file doesn't exist yet —
-  // the modal treats that as "fresh defaults" and creates the file on first
-  // save.
-  ipcMain.handle('settings.readRaw', async () => readNote(settingsPath()));
-
-  // Atomic CAS write to settings.json. Canonicalises through
-  // `globalSettingsSchema` (handled by `writeNote`'s basename dispatch).
-  // Returns the bytes actually written so the caller can keep its CAS
-  // baseline aligned with disk after Zod re-orders keys.
-  ipcMain.handle('settings.writeRaw', async (_, expectedContent: string, newContent: string) =>
-    writeNote(settingsPath(), expectedContent, newContent),
-  );
-
-  ipcMain.handle('project.createNote', async (_, projectPath: string, slug: string) => {
-    await assertUnderConception(projectPath);
-    return createProjectNote(projectPath, slug);
   });
 }
 
@@ -868,23 +286,25 @@ app.whenReady().then(async () => {
   const settings = await readSettings();
   const conceptionPath = settings.lastConceptionPath;
   await setWatchedConception(conceptionPath);
-  lastLayout = settings.layout ?? DEFAULT_LAYOUT;
-  lastRecents = {
+  buildMenu(settings.layout ?? DEFAULT_LAYOUT, {
     paths: settings.recentConceptionPaths ?? [],
     current: conceptionPath,
-  };
-  buildMenu(lastLayout, lastRecents);
+  });
   mainWindow = await createWindow(conceptionPath);
+  setMenuWindow(mainWindow);
   mainWindow.on('closed', () => {
     mainWindow = null;
+    setMenuWindow(null);
   });
 
   app.on('activate', async () => {
     if (BrowserWindow.getAllWindows().length === 0) {
       const settings = await readSettings();
       mainWindow = await createWindow(settings.lastConceptionPath);
+      setMenuWindow(mainWindow);
       mainWindow.on('closed', () => {
         mainWindow = null;
+        setMenuWindow(null);
       });
     }
   });
